Migrate GBDTs (catboost, lightgbm, xgboost) to per-model layout
Apply the TabSTAR Stage 1 recipe to the three GBDT models that previously
relied on the factory loop in `_tabarena_method_metadata_2025_06_12.py`.
Each gets `hpo.py` (search-space generator + `generate_configs_*` helper)
and `info.py` (standalone MethodMetadata + ModelInfo bundle).

The factory file now imports each model's `*_method_metadata` from its
`info.py` and skips them in the loop, so there's a single source of truth
per model and no duplicate registry entries.
