+++ tabarena/tabarena/models/catboost/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.catboost.hpo import gen_catboost
+from tabarena.models.catboost.info import catboost_info, catboost_method_metadata
+
+__all__ = ["catboost_info", "catboost_method_metadata", "gen_catboost"]
