+++ tabarena/tabarena/models/lightgbm/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.lightgbm.hpo import gen_lightgbm
+from tabarena.models.lightgbm.info import lightgbm_info, lightgbm_method_metadata
+
+__all__ = ["gen_lightgbm", "lightgbm_info", "lightgbm_method_metadata"]
