+++ tabarena/tabarena/models/xgboost/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.xgboost.hpo import gen_xgboost
+from tabarena.models.xgboost.info import xgboost_info, xgboost_method_metadata
+
+__all__ = ["gen_xgboost", "xgboost_info", "xgboost_method_metadata"]
